move upgrade class to separate module
